refactor: exception 패키지 생성
