+++ src/test/java/qna/domain/AnswerTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import qna.CannotDeleteException;
+import qna.exception.CannotDeleteException;
 
 import java.util.stream.Stream;
